nullifier reorg
